feature: small cleanup
